backend/requirements.txt
agno
anthropic
fastapi
uvicorn
psycopg2-binary
pyairtable
python-dotenv
sqlalchemy
openai